Squashed 'src/github.com/vektra/tachyon/' changes from 3863494..bd1318d
bd1318d Fix year and name
8f1faa6 Add a LICENSE. Fixes #9
2956faa Merge pull request #8 from skottler/gopath_fixes
1c50495 Fix test:package so it runs to completion
83fdb56 Expose the ability to set ENV vars
de612a9 Add release script
1d1793c Add install option
a2967dd Deal with os/user not working when cross-compiling
e1c9272 Add sample gem module to show yaml modules
16554b8 Merge pull request #7 from skottler/build_deps_task
facb8b5 Add a task to fetch all deps under the build namespace
89ee93e Remove unused variable
3971449 Make Restart and Stop sync by default
d3c5c2c Install atd for testing of upstart
05c139f Include procmgmt modules
8d8845c Add module for interfacing with Upstart
23f36fa Add TaskConfig
5e41367 Fix skipping dbus tests
d8ae979 Flesh out things, add some tests
72b9274 Add package to interface with upstart
d5e0b3d Use tachyon to provision Vagrant
debefbe Pick up go from the calling env
0f3a266 More work on getting travis to run the tests properly
f80e5f9 Add steps to run as root
d89d665 Update apt cache
7cc7691 Only compile against 1.2
b434bcb Add Travis build status
47790a8 Add travis config file
40dca96 Merge pull request #5 from bjoernfan/master
75894a2 goyaml moved their repo and it now only goes by yaml
5cf8383 Fix result formatting
75878ad Add support for referencing a role's specific task file via role::tasks
c4afca1 Fixup perms even if the files match
55f8526 Check for changed and failed in the right places
349177b Report the package name
da1194a Add ability to tag a build as nightly easier
0ef45bf Report error on stderr to not blow up remote invocation
21eaf5f Add ability to only download a file if missing
f691458 Change JSON for a result to be clearer
e54fcbe Refactor FailureResult
ba1de58 Fix error formatting
40f2c55 Gracefully handle refering to an invalid playbook
bf30915 Exit a failed task properly
02f971e Fix copy to refer to files/
cab3d9d Add creates and ignore_failure to shell runners
66f0a6e Allow native modules to get YAML task vars
f71205e Add scripts
6c85958 Fix -t
f8c9f1e Add support for vagrant targets
ed3ee6a Expand vars from a task before invoking
56d1b01 Remove str from Command.Run(), require using parsed args
b5ba9c6 Add inlineMap/Vars, let a Command implement custom arg parsing
cdbcb34 Move Tachyon command to it's own file
96ac7da Fix flowing inner json progress events through
2503d86 Add a json reporter to preserve formating across remote sessions
970e26d Remove broken expansion, fix new way with_items are reported
1982e14 Move remote invocation into a command and use it
4acfa08 map items, explicit ssh persistant ssh
cba3ebb Add s3 module
b74870c Start base roles
36164be Add .gitignore
c825580 Add CacheTime to apt cache update
5e6f259 verify sums in the remote updater
b4def0a Make the remote upload detection based, add enum to Command
fedd6d9 Switch default to be use release
2baaa79 Add support for running a playbook on a remote host
70083ba refactor with_items usage out
53420de Add support for module, late expanded tasks
1713a27 Add role dependencies
a493529 Some refactoring, cleanup
359e679 Use YAML to decode vars
bff0f82 Add tachyon.RunCommand
13aafdc A bit of refactoring
84acec8 Add missing file
d92ea44 Support with_items
24b8bcd Add ability to easily request paths in and outside roles
1b593a9 Refactor paths into common interface
6bd729a Fix script test
9b1d22c Add missing files
cfb6b8c Add script module
8dbcde7 Fix role viewed include paths
a8dc055 Fix aux tests
fb7161c Add missing test fixtures
7d1b4ca Handle inline vars on role inclusion
a614cb2 Add role variables
8bd18b5 Add handlers and vars to roles
2672016 Reduce the future tests to 1s instead of 5s
8bd0f17 Begin adding role support
47b768c Inject additional keys in task include
850b540 Add vars hash support to included tasks
f165d47 Add task include variables
958b239 Add timing and file indicators to Task
a2cb0e1 Handle StdoutPipe properly
305ffe5 Fix apt
07d2e7c Add fix command/shell output, add RunCapture
ffbf921 Move all printing into reporter
5b6e1a1 Cleanup play parsing
fba4dd4 Add included tasks
f517088 Bunch of stuff I missed in the last commit
329d9d9 download, stdout trimming
22029ea Add register declaration
bc4d42e Expose Maps via template expansions
14b7743 Lace futures into lisp values
69370a3 Get futures working
feab6ed Start work on futures
140cd77 Move apt to package/apt
9357f36 Get apt working nicely
8f3c390 Start the apt module
96aa5e0 Tests for the current builtins
610f1ee Add Result data
24db9f1 A bit of error checking
e6aa18b Actually replace PlayEnv with Runner
0bb3cf1 Remove PlayEnv, add Runner
a607421 Add a var scope to included playbooks
6f8e97e Add (or) special form and change quote char to :
8ae5d0b Refactor vars access behind Scope interface
d29c939 Begin to introduce more nested environments
ee79832 Add options and nested envs
88f6a25 Add support for $vars
e1a3073 Add support for includes

git-subtree-dir: src/github.com/vektra/tachyon
git-subtree-split: bd1318dc8a2457a00120239dd7f81b93a816cc00

+++ lisp/cons_test.go
@@ -5,9 +5,9 @@ import (
 )
 
 func cons() Cons {
-	v1 := &Value{numberValue, 1.0}
-	v2 := &Value{numberValue, 2.0}
-	v3 := &Value{numberValue, 3.0}
+	v1 := &Value{numberValue, int64(1)}
+	v2 := &Value{numberValue, int64(2)}
+	v3 := &Value{numberValue, int64(3)}
 	c2 := &Value{consValue, &Cons{v3, &Value{nilValue, nil}}}
 	c1 := &Value{consValue, &Cons{v2, c2}}
 	return Cons{v1, c1}
@@ -15,9 +15,9 @@ func cons() Cons {
 
 func TestConsMap(t *testing.T) {
 	s, _ := cons().Map(func(v Value) (Value, error) {
-		return Value{numberValue, v.val.(float64) + 1}, nil
+		return Value{numberValue, v.val.(int64) + 1}, nil
 	})
-	if len(s) != 3 || s[0].val != 2.0 || s[1].val != 3.0 || s[2].val != 4.0 {
+	if len(s) != 3 || s[0].val != int64(2) || s[1].val != int64(3) || s[2].val != int64(4) {
 		t.Errorf("Expected (1 2 3), got %v", s)
 	}
 }
@@ -31,7 +31,7 @@ func TestConsLen(t *testing.T) {
 
 func TestConsVector(t *testing.T) {
 	s := cons().Vector()
-	if len(s) != 3 || s[0].val != 1.0 || s[1].val != 2.0 || s[2].val != 3.0 {
+	if len(s) != 3 || s[0].val != int64(1) || s[1].val != int64(2) || s[2].val != int64(3) {
 		t.Errorf("Expected (1 2 3), got %v", s)
 	}
 }
